fix: excluded fields
This list is generated by querying Zuora with all fields, and removing
the faulty fields, one by one, until the query succeeds.

Until recently, Zuora was returning HTTP 500 Internal Server Error when
a non-queryable field was passed in the query. It now returns HTTP 400
Bad Request Error, but this exception was not rescued in the service
class generating the list of excluded fields.