Fetch error class using faultcode
Some error response hash now contains "faultcode",
for example the body could be:

{
  "detail" => {
    "MalformedQueryFault" => {
      "FaultMessage" => "invalid field for query: Account.foo",
      "FaultCode" => "INVALID_FIELD" }
  },
  "faultcode" => "fns:INVALID_FIELD",
  "faultstring" => "invalid field for query: Account.foo"
}

So when response is a hash, we try to find faultcode to lookup
error class instead of body text matching.

Body text matching is bad because it can cause false positives,
when data happens to contain error code in the custom fields.

Bulk operation (which returns a hash) is the exception.
We assume it does not contain custom fields info based on experience.